+++ CHANGELOG.md
@@ -13,6 +13,7 @@ is silenced.
 - Added --disable-assets flag to sensu-agent.
 - Added ability to query mutators to the GraphQL service
 - Added ability to query event filters to the GraphQL service
+- Added prometheus metrics for topics in wizard bus and agent sessions.
 - The buffer size and worker count of keepalived, eventd & pipelined can now be
 configured on sensu-backend.
 
@@ -23,6 +24,7 @@ POST & PUT requests instead of `204 No Content`.
 ### Fixed
 - The REST API now returns an error when trying to delete an entity that does
 not exist.
+- Fixed a bug where basic authorization was not being performed on the agent websocket connection.
 - Fixed an aliasing regression where event timestamps from the /events API
 were not getting properly populated.